whatsapp css and move to feed
