+++ web/src/elements/Btn/index.tsx
@@ -113,7 +113,7 @@ export default function Btn({
 
     switch (btnType) {
         case BtnType.corporative:
-            classNames.push("btn ");
+            classNames.push("btn btn--corporative");
             break;
         case BtnType.submit:
             classNames.push("btn btn--black");
