+++ web/src/layouts/Feed/index.tsx
@@ -8,7 +8,9 @@ import Btn, {
 import t from 'i18n';
 import {
   IoAdd,
+  IoCallOutline,
   IoCloseOutline,
+  IoLogoWhatsapp,
   IoPersonOutline,
 } from 'react-icons/io5';
 
@@ -23,7 +25,7 @@ import {
   DeletePost,
   LoadPosts,
 } from 'state/Posts';
-import { isAdmin } from 'state/Users';
+import { GetPhone, isAdmin } from 'state/Users';
 import { useStore } from 'state';
 import MessageNew from 'components/feed/MessageNew';
 import { PrivacyType } from 'shared/types/privacy.enum';
@@ -69,6 +71,8 @@ export default function Feed({ button }: { button: Button }) {
   return (
     <div className="feed-container">
       <div className="card-button__actions">
+        <ShowPhone user={button.owner} />
+
         <>
           {sessionUser && isButtonOwner && (
             <>
@@ -134,6 +138,78 @@ export default function Feed({ button }: { button: Button }) {
     </div>
   );
 }
+
+export function ShowPhone({ user }) {
+  const [showPhone, toggleShowPhone] = useState(false);
+  const [phone, setPhone] = useState(null);
+
+  useEffect(() => {
+    toggleShowPhone(() => true)
+  }, [phone])
+  const jumpTo = (url) => {
+    window.location.replace(url);
+  }
+  const onCallClick = () => {
+    if (phone == null) {
+      store.emit(
+        new GetPhone(
+          user.id,
+          (phone) => {
+            setPhone(phone);
+            jumpTo(`tel:${phone}`);
+          },
+          () => {},
+        ),
+      );
+    }else{
+      jumpTo(`tel:${phone}`);
+    }
+  };
+
+  const onWassapClick = ( ) => {
+    if (phone == null) {
+      store.emit(
+        new GetPhone(
+          user.id,
+          (phone) => {
+            setPhone(phone);
+            jumpTo(`whatsapp://send?phone=+${phone}`)
+          },
+          () => {},
+        ),
+      );
+    }else{
+      jumpTo(`whatsapp://send?phone=+${phone}`)
+    }
+  }
+
+  return (
+    <>
+      {user?.publishPhone && (
+        <>
+          <Btn
+            btnType={BtnType.corporative}
+            contentAlignment={ContentAlignment.center}
+            iconLeft={IconType.circle}
+            iconLink={<IoCallOutline />}
+            onClick={() => onCallClick()}
+          />
+          {showPhone && <>{phone}</>}
+          {user.showWassap && 
+            <Btn
+              btnType={BtnType.corporative}
+              contentAlignment={ContentAlignment.center}
+              iconLeft={IconType.circle}
+              iconLink={<IoLogoWhatsapp />}
+              onClick={() => onWassapClick()}
+            />
+          }
+        </>
+      )}
+    </>
+  );
+}
+
 export function FeedElement({
   post,
   sessionUser,
