+++ web/src/pages/ProfileEdit/index.tsx
@@ -219,12 +219,7 @@ export default function ProfileEdit() {
                         explain={t('user.adminPhonePublishExplain')}
                         onChanged={(value) => {setValue('publishPhone', value)}}
                       />
-                      {/* <FieldCheckbox
-                        name='showWhatsapp'
-                        defaultValue={sessionUser.showWhatsapp}
-                        text={t('user.showWhatsapp')}
-                        onChanged={(value) => {setValue('showWhatsapp', value)}}
-                      /> */}
+               
                     </>
                   }
                   <FieldCheckbox
